backend/requirements.txt
fastapi
uvicorn
motor
beanie
google-generativeai
python-dotenv
pydantic
numpy
scikit-learn
ollama